server/requirements.txt
python-telegram-bot==20.6
boto3==1.28.57
pydub==0.25.1
python-docx==0.8.11